move compat structs to separate modules
